+++ example/lib/src/home_binding.dart
@@ -2,14 +2,23 @@ import 'package:easy_binding/easy_binding.dart';
 import 'package:example/src/home_controller.dart';
 import 'package:get_it/get_it.dart';
 
+/// Binding for the Home page.
+///
+/// This binding is responsible for registering and unregistering
+/// the [HomeController] in the GetIt dependency injection container.
+///
+/// The controller will be automatically registered when the Home route
+/// is created and unregistered when the route is disposed.
 class HomeBinding extends EasyBinding {
   @override
-  Future<void> onInit() async {
+  void onInit() {
+    // Register the HomeController as a singleton in GetIt
     GetIt.instance.registerSingleton<HomeController>(HomeController());
   }
 
   @override
-  Future<void> onDispose() async {
+  void onDispose() {
+    // Unregister the HomeController from GetIt
     GetIt.instance.unregister<HomeController>();
   }
 }